% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidnapr.R
\docType{package}
\name{kidnapr}
\alias{kidnapr}
\alias{kidnapr-package}
\title{kidnapr: Shiny application to collect data for future inclusion in the
kidstats package.}
\usage{
kidnapr()
}
\description{
\code{kidnapr} Purpose built Shiny application to collect data.

KidnapR shiny app call
}
\examples{
kidnapr()
}

