% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KScollect.R
\docType{package}
\name{KScollect}
\alias{KScollect}
\alias{KScollect-package}
\title{KScollect: Shiny application to collect data for future inclusion in the
KidStats package.}
\usage{
KScollect()
}
\description{
\code{KScollect} Purpose built Shiny application to collect data.

KScollect shiny app call
}
\examples{
KScollect()
}

